transcricao_de_audio/requirements.txt
openai-whisper==20240930
torch==2.6.0
pydub==0.25.1
pandas==2.2.3

